Renamed mBookings to getMyBookings.js
