+++ server/models/getMyBookings.js
@@ -1,7 +1,8 @@
 const client = require('../../db/index.js').client;
 
-const getMyBookings = () => {
-
+const getMyBookings = (id) => {
+  let query = `SELECT * FROM bookings WHERE renter_id=$1 ORDER BY booking_id`;
+  let values = [id]
 }
 
 module.exports.getMyBookings = getMyBookings;
